+++ reductstore/src/ext/ext_repository/load.rs
@@ -13,6 +13,8 @@ use std::path::PathBuf;
 use std::sync::Arc;
 use tokio::sync::RwLock as AsyncRwLock;
 
+const EXTENSION_VERSION: &str = "0.2.3";
+
 impl ExtRepository {
     pub(super) fn try_load(
         paths: Vec<PathBuf>,
@@ -94,7 +96,7 @@ mod tests {
             info,
             IoExtensionInfo::builder()
                 .name("test-ext")
-                .version("0.1.1")
+                .version(EXTENSION_VERSION)
                 .build()
         );
     }
@@ -128,19 +130,30 @@ mod tests {
     #[fixture]
     fn ext_repo(ext_settings: ExtSettings) -> ExtRepository {
         // This is the path to the build directory of the extension from ext_stub crate
-        const EXTENSION_VERSION: &str = "0.1.1";
-
-        if !cfg!(target_arch = "x86_64") {
-            panic!("Unsupported architecture");
-        }
 
         let file_name = if cfg!(target_os = "linux") {
             // This is the path to the build directory of the extension from ext_stub crate
-            "libtest_ext-x86_64-unknown-linux-gnu.so"
+            if cfg!(target_arch = "aarch64") {
+                "libtest_ext-aarch64-unknown-linux-gnu.so"
+            } else if cfg!(target_arch = "x86_64") {
+                "libtest_ext-x86_64-unknown-linux-gnu.so"
+            } else {
+                panic!("Unsupported architecture")
+            }
         } else if cfg!(target_os = "macos") {
-            "libtest_ext-x86_64-apple-darwin.dylib"
+            if cfg!(target_arch = "aarch64") {
+                "libtest_ext-aarch64-apple-darwin.dylib"
+            } else if cfg!(target_arch = "x86_64") {
+                "libtest_ext-x86_64-apple-darwin.dylib"
+            } else {
+                panic!("Unsupported architecture")
+            }
         } else if cfg!(target_os = "windows") {
-            "libtest_ext-x86_64-pc-windows-gnu.dll"
+            if cfg!(target_arch = "x86_64") {
+                "libtest_ext-x86_64-pc-windows-gnu.dll"
+            } else {
+                panic!("Unsupported architecture")
+            }
         } else {
             panic!("Unsupported platform")
         };