+++ CHANGELOG.md
@@ -17,11 +17,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add "@" prefix to computed labels, [PR-815](https://github.com/reductstore/reductstore/pull/815)
 - Refactor Extension API for multi-line CSV processing, ReductSelect v0.2.0, [PR-823](https://github.com/reductstore/reductstore/pull/823)
 
+### Fixed
+
+- Fix hanging query request if no extension registered, [PR-830](https://github.com/reductstore/reductstore/pull/830)
+
 ## [1.15.3] - 2025-05-26
 
 ## Fixed
 
 - Fix lock of write channel for small chunks, [PR-834](https://github.com/reductstore/reductstore/pull/834)
+
 
 ## [1.15.2] - 2025-05-21